Refactor src/api/interfaces.ts into types and transforms
- Split `src/api/interfaces.ts` into `src/api/types.ts` and `src/api/transforms.ts`.
- Update `src/api/interfaces.ts` to re-export from the new files for backward compatibility.
- Update `src/api/item.ts` to import types from `./types.ts` to resolve circular dependency with `src/api/index.ts`.
- Update `src/api/stories.ts` to use new imports.
- Mark tasks as completed in `PREVENT_CONFLICTS.md`.
